Parameter validation
In order to improve robustness I have some parameter handling now.
If we receive no parameters we default to the default IP and file path.
If we receive 1 parameter:
We check if its a directory; If so we use this and default port
We check if it is a port number; If so we use this and default filepath